adding corrections
