+++ src/Blocks/custom/form-selector/manifest.json
@@ -125,7 +125,49 @@
 		{
 			"label": "Jira",
 			"slug": "jira",
-			"blockName": "eightshift-forms/jira"
+			"blockName": "eightshift-forms/jira",
+			"innerBlocks": [
+				[
+					"eightshift-forms/input",
+					{
+						"inputInputFieldLabel": "E-mail",
+						"inputInputType": "email",
+						"inputInputName": "email",
+						"inputInputIsRequired": true,
+						"inputInputIsEmail": true,
+						"inputInputDisabledOptions": [
+							"inputInputIsRequired",
+							"inputInputName"
+						]
+					}
+				],
+				[
+					"eightshift-forms/input",
+					{
+						"inputInputFieldLabel": "Name",
+						"inputInputIsRequired": true,
+						"inputInputName": "name"
+					}
+				],
+				[
+					"eightshift-forms/input",
+					{
+						"inputInputFieldLabel": "Subject",
+						"inputInputIsRequired": true,
+						"inputInputName": "subject"
+					}
+				],
+				[
+					"eightshift-forms/textarea",
+					{
+						"textareaTextareaFieldLabel": "Message",
+						"textareaTextareaName": "message"
+					}
+				],
+				[
+					"eightshift-forms/submit"
+				]
+			]
 		}
 	]
 }
