+++ src/Integrations/Hubspot/Hubspot.php
@@ -513,13 +513,11 @@ class Hubspot extends AbstractFormBuilder implements MapperInterface, ServiceInt
 								'checkboxLabel' => $communicationLabel,
 								'checkboxValue' => \wp_strip_all_tags($communicationLabel),
 								'checkboxIsChecked' => $communicationIsHidden,
-								'checkboxIsDisabled' => $communicationIsHidden,
 								'checkboxHideLabel' => $communicationIsHidden,
 								'checkboxDisabledOptions' => $this->prepareDisabledOptions('checkbox', [
 									'checkboxLabel',
 									'checkboxValue',
 									'checkboxIsChecked',
-									'checkboxIsDisabled',
 									'checkboxHideLabel',
 								], false),
 							],
@@ -552,13 +550,11 @@ class Hubspot extends AbstractFormBuilder implements MapperInterface, ServiceInt
 							'checkboxLabel' => $processingLabel,
 							'checkboxValue' => \wp_strip_all_tags($processingLabel),
 							'checkboxIsChecked' => $processingIsHidden,
-							'checkboxIsDisabled' => $processingIsHidden,
 							'checkboxHideLabel' => $processingIsHidden,
 							'checkboxDisabledOptions' => $this->prepareDisabledOptions('checkbox', [
 								'checkboxLabel',
 								'checkboxValue',
 								'checkboxIsChecked',
-								'checkboxIsDisabled',
 								'checkboxHideLabel',
 							], false),
 						],
@@ -589,14 +585,12 @@ class Hubspot extends AbstractFormBuilder implements MapperInterface, ServiceInt
 								'component' => 'checkbox',
 								'checkboxLabel' => HubspotClient::HUBSPOT_CONSENT_LEGITIMATE,
 								'checkboxIsChecked' => true,
-								'checkboxIsDisabled' => true,
 								'checkboxHideLabel' => true,
 								'checkboxValue' => 'true',
 								'checkboxDisabledOptions' => $this->prepareDisabledOptions('checkbox', [
 									'checkboxLabel',
 									'checkboxValue',
 									'checkboxIsChecked',
-									'checkboxIsDisabled',
 									'checkboxHideLabel',
 								], false),
 							],
